refactor(*): organize constants more specific
- platform
- discord
- spa
