+++ src/main/java/gdsc/konkuk/platformcore/global/consts/DiscordConstants.java
@@ -1,0 +1,9 @@
+package gdsc.konkuk.platformcore.global.consts;
+
+public class DiscordConstants {
+
+    public static final String DISCORD_ERROR_TITLE = "\uD83E\uDDE8 치명적인 서버 에러 발생!";
+    public static final String DISCORD_ERROR_DESCRIPTION =
+            "\uD83C\uDD98 서버에서 치명적인 에러가 발생했습니다. 빠르게 확인해주세요. \uD83C\uDD98";
+    public static final String DISCORD_ERROR_TIME_TEXT = "\uD83D\uDD50 에러 발생 시간 : ";
+}
